pyproject.toml
[tool.poetry]
name = "hr-copilot-v2"
version = "0.1.0"
description = "AI-powered HR resume screening assistant system"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.8.1"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.0"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.23"
alembic = "^1.13.0"
python-multipart = "^0.0.6"
python-docx = "^1.1.0"
pypdf2 = "^3.0.1"
litellm = "^1.17.0"
aiofiles = "^23.2.0"
email-validator = "^2.2.0"
httpx = {extras = ["socks"], version = "^0.28.1"}
reportlab = "^4.4.1"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
